requirements.txt
pandas
SQLAlchemy
psycopg2-binary
python-dateutil
python-dotenv
requests
streamlit
